@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:soundboard/core/services/innebandy_api/domain/entities/match.dart';
 import 'package:soundboard/features/screen_home/presentation/live/data/class_match_event_type.dart';
 import 'class_ssml_goalevent.dart';
 import 'class_ssml_penaltyevent.dart';
-import 'package:soundboard/features/innebandy_api/data/class_match.dart';
 
 class EventCardSsml {
   final dynamic data;
