@@ -2,15 +2,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
-import 'package:soundboard/common_widgets/widget_match.dart';
-import 'package:soundboard/constants/default_constants.dart';
-import 'package:soundboard/features/innebandy_api/data/class_match.dart';
-import 'package:soundboard/features/screen_home/presentation/board/classes/class_2_RowPlayerPresentation.dart';
 import 'package:soundboard/features/screen_home/presentation/board/classes/class_0_player_progress_bar.dart';
-import 'package:soundboard/features/screen_home/presentation/board/classes/class_3_ratata.dart';
-import 'package:soundboard/features/screen_home/presentation/board/classes/class_4_lineup.dart';
-import 'package:soundboard/features/screen_home/presentation/board/classes/class_5_timeout.dart';
+
 import 'package:soundboard/features/screen_home/presentation/board/classes/class_1_stop_goal_row.dart';
+import 'package:soundboard/features/screen_home/presentation/board/classes/class_jingle_grid_section.dart';
+import 'package:soundboard/features/screen_home/presentation/board/classes/class_horizontal_volume_control.dart';
 
 class BoardSection extends ConsumerWidget {
   final double width;
@@ -19,36 +15,24 @@ class BoardSection extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final selectedMatch = ref.watch(selectedMatchProvider);
-
     return SizedBox(
       width: width,
-      child: Padding(
-        padding: const EdgeInsets.all(5.0),
+      child: const Padding(
+        padding: EdgeInsets.all(5.0),
         child: Align(
           alignment: Alignment.topCenter,
           child: SingleChildScrollView(
             child: Column(
               children: [
-                SizedBox(
-                  width: ScreenSizeUtil.getWidth(
-                    context,
-                    maxWidth: ScreenSizeUtil.getSoundboardSize(context),
-                  ),
-                  child: MatchButton2(match: selectedMatch, readonly: true),
-                ),
-                const PlayerProgressBar(),
-                const Gap(10),
-                const StopGoalRow(),
-                const Gap(10),
-                const RowPlayerPresentation(),
-                const Gap(10),
-                const Row1Ratata(),
-                const Gap(10),
-                const Row2lineup(),
-                const Gap(10),
-                const Row3timeout(),
-                const Gap(10),
+                HorizontalVolumeControl(),
+                Gap(10),
+                PlayerProgressBar(),
+                Gap(10),
+                StopGoalRow(),
+                Gap(10),
+                // RowPlayerPresentation(),
+                // const Gap(10),
+                JingleGridSection(),
               ],
             ),
           ),
