@@ -1,10 +1,10 @@
 // First, create a separate widget for the period scores to better manage the state
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:soundboard/features/innebandy_api/data/class_match.dart';
+import 'package:soundboard/core/services/innebandy_api/domain/entities/match.dart';
 import 'package:soundboard/features/screen_home/application/audioplayer/ssml/class_ssml_periodevent.dart';
 import 'package:soundboard/features/screen_home/presentation/events/classes/class_divider.dart';
-import 'package:soundboard/utils/logger.dart';
+import 'package:soundboard/core/utils/logger.dart';
 import 'package:collection/collection.dart';
 
 class PeriodScores extends ConsumerWidget {
@@ -32,7 +32,6 @@ class PeriodScores extends ConsumerWidget {
                 style: TextStyle(
                   fontSize: 12,
                   fontStyle: FontStyle.italic,
-                  // fontWeight: FontWeight.bold,
                   color: Theme.of(context).colorScheme.onPrimaryContainer,
                 ),
                 "${selectedMatch.goalsHomeTeam} - ${selectedMatch.goalsAwayTeam}",
@@ -110,8 +109,8 @@ class PeriodScores extends ConsumerWidget {
 
     return Expanded(
       child: TextButton(
-        onPressed:
-            () => SsmlPeriodEvent(period: period, ref: ref).getSay(context),
+        onPressed: () =>
+            SsmlPeriodEvent(period: period, ref: ref).getSay(context),
         child: Column(
           children: [
             Text(
@@ -130,7 +129,6 @@ class PeriodScores extends ConsumerWidget {
                 fontStyle: FontStyle.italic,
                 color: Theme.of(context).colorScheme.onPrimaryContainer,
               ),
-
               getPeriodScore(period),
             ),
           ],
