@@ -5,8 +5,12 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:soundboard/features/screen_home/presentation/board/board.dart';
 import 'package:soundboard/features/screen_home/presentation/events/events.dart';
 import 'package:soundboard/features/screen_home/presentation/lineup/lineup.dart';
-import 'package:soundboard/features/screen_home/presentation/volume/volume.dart';
 
+/// Main screen that displays the soundboard interface.
+///
+/// The layout is responsive and adapts to different screen sizes:
+/// - On Android: Uses a single column layout with full width
+/// - On Windows: Uses a multi-column layout with dynamic widths
 class HomeScreen extends ConsumerStatefulWidget {
   const HomeScreen({super.key});
 
@@ -19,6 +23,7 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
 
   @override
   void dispose() {
+    scrollController.dispose();
     super.dispose();
   }
 
@@ -37,24 +42,57 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
             if (Platform.isAndroid) {
               return BoardSection(width: constraints.maxWidth);
             }
+
+            // Calculate widths for Windows layout
+            final totalWidth = constraints.maxWidth;
+
+            // Define fixed maximum widths for each section
+            const maxBoardWidth = 400.0;
+            const maxVolumeWidth = 120.0;
+            const maxEventsWidth = 350.0;
+
+            // Define minimum widths for each section
+            const minBoardWidth = 280.0;
+            const minVolumeWidth = 80.0;
+            const minEventsWidth = 240.0;
+            const minLineupWidth = 280.0;
+
+            // Calculate widths for the first three sections
+            final boardWidth =
+                totalWidth *
+                0.3.clamp(
+                  minBoardWidth / totalWidth,
+                  maxBoardWidth / totalWidth,
+                );
+            final volumeWidth =
+                totalWidth *
+                0.1.clamp(
+                  minVolumeWidth / totalWidth,
+                  maxVolumeWidth / totalWidth,
+                );
+            final eventsWidth =
+                totalWidth *
+                0.25.clamp(
+                  minEventsWidth / totalWidth,
+                  maxEventsWidth / totalWidth,
+                );
+
+            // Calculate remaining width for lineup section
+            final usedWidth = boardWidth + volumeWidth + eventsWidth;
+            final lineupWidth = (totalWidth - usedWidth).clamp(
+              minLineupWidth,
+              double.infinity,
+            );
+
             return Row(
               children: [
-                const BoardSection(width: 400),
-                if (Platform.isWindows) ...[
-                  const VolumeSection(
-                    width: 100,
-                  ),
-                  EventsSection(
-                    scrollController: scrollController,
-                    width: 350,
-                  ),
-                  VerticalDivider(
-                    thickness: 1.0,
-                    width: 2.0,
-                    color: Theme.of(context).colorScheme.onSurface,
-                  ),
-                  const LineupSection(), // 502 = 400 + 100 + 2
-                ],
+                BoardSection(width: boardWidth),
+                // VolumeSection(width: volumeWidth),
+                EventsSection(
+                  scrollController: scrollController,
+                  width: eventsWidth,
+                ),
+                LineupSection(width: lineupWidth),
               ],
             );
           },
