@@ -1,20 +1,22 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:soundboard/constants/globals.dart';
-import 'package:soundboard/constants/providers.dart';
-import 'package:soundboard/features/cloud_text_to_speech/providers.dart';
-import 'package:soundboard/features/innebandy_api/data/class_match.dart';
-import 'package:soundboard/features/jingle_manager/application/class_audiocategory.dart';
+import 'package:soundboard/core/services/jingle_manager/jingle_manager_provider.dart';
+import 'package:soundboard/core/utils/providers.dart';
+import 'package:soundboard/core/services/cloud_text_to_speech/providers.dart';
+import 'package:soundboard/core/services/innebandy_api/domain/entities/match.dart';
+
+import 'package:soundboard/core/services/jingle_manager/class_audiocategory.dart';
 import 'package:soundboard/features/screen_home/application/audioplayer/data/class_audiomanager.dart';
+import 'package:soundboard/features/screen_home/application/audioplayer/data/class_audio.dart';
 import 'package:soundboard/features/screen_home/presentation/lineup/classes/class_lineup_data.dart';
 import 'package:soundboard/features/screen_home/presentation/lineup/classes/class_new_notepad.dart';
-import 'package:soundboard/properties.dart';
-import 'package:soundboard/utils/logger.dart';
+import 'package:soundboard/features/screen_home/presentation/lineup/providers/manual_lineup_providers.dart';
+import 'package:soundboard/features/screen_home/presentation/lineup/widgets/manual_lineup_entry_widget.dart';
+import 'package:soundboard/features/screen_home/presentation/lineup/widgets/manual_event_generator_widget.dart';
+import 'package:soundboard/core/properties.dart';
+import 'package:soundboard/core/utils/logger.dart';
 // [Keep other imports...]
-
-// Add the loading state provider
-final isLoadingProvider = StateProvider<bool>((ref) => false);
 
 class Lineup extends ConsumerStatefulWidget {
   final double availableWidth;
@@ -27,123 +29,90 @@ class Lineup extends ConsumerStatefulWidget {
 }
 
 class _LineupState extends ConsumerState<Lineup> {
-  final List<String> loadingMessages = [
-    "Warming up the virtual vocal cords...",
-    "Teaching robots to sing...",
-    "Composing your audio masterpiece...",
-    "Converting text to sweet melodies...",
-    "Preparing your lineup announcement...",
-    "Tuning the digital microphone...",
-    "Getting the virtual crowd excited...",
-    "Clearing throat (beep boop)...",
-    "Loading commentary powers...",
-    "Summoning the sports announcer spirit...",
-  ];
   final Logger logger = const Logger('Lineup');
-
-  int currentMessageIndex = 0;
-  Timer? messageRotationTimer;
-
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    messageRotationTimer?.cancel();
-    super.dispose();
-  }
-
-  void startMessageRotation() {
-    currentMessageIndex = 0;
-    messageRotationTimer?.cancel();
-    messageRotationTimer =
-        Timer.periodic(const Duration(milliseconds: 2000), (timer) {
-      setState(() {
-        currentMessageIndex =
-            (currentMessageIndex + 1) % loadingMessages.length;
-      });
-    });
-  }
-
-  void stopMessageRotation() {
-    messageRotationTimer?.cancel();
-    messageRotationTimer = null;
-  }
 
   @override
   Widget build(BuildContext context) {
     final selectedMatch = ref.watch(selectedMatchProvider);
-    final isLoading = ref.watch(isLoadingProvider);
+    final isManualMode = ref.watch(isManualLineupModeProvider);
     final theme = Theme.of(context);
 
-    return Stack(
-      children: [
-        SizedBox(
-          width: widget.availableWidth,
-          height: widget.availableHeight,
-          child: Padding(
-            padding: const EdgeInsets.all(5.0),
-            child: Column(
-              children: [
-                _buildHeader(theme, selectedMatch),
-                _buildDivider(theme),
-                LineupData(
-                    availableWidth: widget.availableWidth,
-                    availableHeight: widget.availableHeight),
-              ],
-            ),
-          ),
-        ),
-        if (isLoading)
-          Container(
-            width: widget.availableWidth,
-            height: widget.availableHeight,
-            color: Colors.black54,
-            child: Center(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
+    return SizedBox(
+      width: widget.availableWidth,
+      height: widget.availableHeight,
+      child: Padding(
+        padding: const EdgeInsets.all(5.0),
+        child: Column(
+          children: [
+            // Mode Toggle Switch
+            Container(
+              padding: const EdgeInsets.symmetric(
+                horizontal: 8.0,
+                vertical: 4.0,
+              ),
+              decoration: BoxDecoration(
+                color: theme.colorScheme.surfaceContainer,
+                borderRadius: BorderRadius.circular(8.0),
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
-                  const CircularProgressIndicator(
-                    valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-                  ),
-                  const SizedBox(height: 16),
-                  AnimatedSwitcher(
-                    duration: const Duration(milliseconds: 500),
-                    transitionBuilder:
-                        (Widget child, Animation<double> animation) {
-                      return FadeTransition(
-                        opacity: animation,
-                        child: SlideTransition(
-                          position: Tween<Offset>(
-                            begin: const Offset(0.0, 0.5),
-                            end: Offset.zero,
-                          ).animate(animation),
-                          child: child,
-                        ),
-                      );
+                  const Text('Manual Mode'),
+                  Switch(
+                    value: isManualMode,
+                    onChanged: (value) {
+                      ref.read(isManualLineupModeProvider.notifier).state =
+                          value;
                     },
-                    child: Text(
-                      loadingMessages[currentMessageIndex],
-                      key: ValueKey<int>(currentMessageIndex),
-                      style: const TextStyle(
-                        color: Colors.white,
-                        fontSize: 16,
-                        fontWeight: FontWeight.w500,
-                      ),
-                      textAlign: TextAlign.center,
-                    ),
                   ),
                 ],
               ),
             ),
-          ),
-      ],
+            const SizedBox(height: 8),
+
+            // Only show header section in API mode (not manual mode)
+            if (!isManualMode) ...[
+              Expanded(
+                flex: 1,
+                child: SingleChildScrollView(
+                  child: _buildHeader(theme, selectedMatch),
+                ),
+              ),
+              _buildDivider(theme),
+            ],
+            Expanded(
+              flex: 3,
+              child: isManualMode
+                  ? ManualLineupEntryWidget(
+                      availableWidth: widget.availableWidth,
+                      availableHeight: widget.availableHeight * 0.6,
+                    )
+                  : LineupData(
+                      availableWidth: widget.availableWidth,
+                      availableHeight: widget.availableHeight,
+                    ),
+            ),
+
+            // Add Manual Event Generator at the bottom when in manual mode
+            if (isManualMode) ...[
+              _buildDivider(theme),
+              Expanded(
+                flex: 2,
+                child: ManualEventGeneratorWidget(
+                  availableWidth: widget.availableWidth,
+                  availableHeight: widget.availableHeight * 0.4,
+                ),
+              ),
+            ],
+          ],
+        ),
+      ),
     );
   }
 
   Widget _buildHeader(ThemeData theme, dynamic selectedMatch) {
+    final isManualMode = ref.watch(isManualLineupModeProvider);
+
     return Container(
       decoration: BoxDecoration(
         color: theme.colorScheme.primaryContainer,
@@ -153,14 +122,25 @@ class _LineupState extends ConsumerState<Lineup> {
       child: Column(
         children: [
           _buildPlayButton(theme, selectedMatch),
-          _buildDivider(theme),
-          const Row(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Expanded(child: GoalInputWidget(team: "homeTeam")),
-              Expanded(child: GoalInputWidget(team: "awayTeam")),
-            ],
-          ),
+          // Only show the goal/penalty inputs in API mode (not manual mode)
+          if (!isManualMode) ...[
+            _buildDivider(theme),
+            const Row(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Expanded(child: GoalInputWidget(team: "homeTeam")),
+                Expanded(child: GoalInputWidget(team: "awayTeam")),
+              ],
+            ),
+            _buildDivider(theme),
+            const Row(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Expanded(child: PenaltyInputWidget(team: "homeTeam")),
+                Expanded(child: PenaltyInputWidget(team: "awayTeam")),
+              ],
+            ),
+          ],
         ],
       ),
     );
@@ -169,6 +149,10 @@ class _LineupState extends ConsumerState<Lineup> {
   Widget _buildPlayButton(ThemeData theme, dynamic selectedMatch) {
     return TextButton(
       onPressed: () => _handlePlayLineup(selectedMatch),
+      onLongPress: () {
+        final text = _getLineupText(selectedMatch);
+        _showTextDialog(context, text);
+      },
       child: Text(
         "Lineup (Click to Play)",
         style: TextStyle(
@@ -180,37 +164,115 @@ class _LineupState extends ConsumerState<Lineup> {
     );
   }
 
-  Future<void> _handlePlayLineup(dynamic selectedMatch) async {
-    // Show loading indicator and start message rotation
-    ref.read(isLoadingProvider.notifier).state = true;
-    startMessageRotation();
+  String _getLineupText(dynamic selectedMatch) {
+    final introText = _stripSsmlTags(selectedMatch.introSsml(ref));
+    final homeTeamText = _stripSsmlTags(selectedMatch.homeTeamSsml(ref));
+    final awayTeamText = _stripSsmlTags(selectedMatch.awayTeamSsml(ref));
 
+    return "$introText\n\n$awayTeamText\n\n$homeTeamText";
+  }
+
+  String _stripSsmlTags(String text) {
+    // Remove SSML/XML tags using regex
+    return text.replaceAll(RegExp(r'<[^>]*>'), '');
+  }
+
+  void _showTextDialog(BuildContext context, String text) {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Lineup Announcement Text'),
+        content: SingleChildScrollView(child: Text(text)),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('Close'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Future<void> _handlePlayLineup(dynamic selectedMatch) async {
     try {
+      // Debug: Check the lineup data before generating SSML
+      final lineupData = ref.read(effectiveLineupProvider);
+      logger.d("[_handlePlayLineup] Lineup matchId: ${lineupData.matchId}");
+      logger.d(
+        "[_handlePlayLineup] Home team players count: ${lineupData.homeTeamPlayers.length}",
+      );
+      logger.d(
+        "[_handlePlayLineup] Away team players count: ${lineupData.awayTeamPlayers.length}",
+      );
+
+      // Get the jingle manager from provider
+      final jingleManagerAsync = ref.read(jingleManagerProvider);
+      final jingleManager = await jingleManagerAsync.when(
+        data: (manager) => manager,
+        loading: () => throw Exception('JingleManager not loaded'),
+        error: (error, stack) => throw Exception('JingleManager error: $error'),
+      );
+
       final textToSpeechService = ref.read(textToSpeechServiceProvider);
       // final speech =
       //     await textToSpeechService.getTtsNoFile(text: selectedMatch.ssml);
 
-      final welcomeTTS =
-          await textToSpeechService.getTtsNoFile(text: selectedMatch.introSsml);
+      final welcomeTTS = await textToSpeechService.getTtsNoFile(
+        text: selectedMatch.introSsml(ref),
+      );
       final homeTeamTTS = await textToSpeechService.getTtsNoFile(
-          text: selectedMatch.homeTeamSsml);
+        text: selectedMatch.homeTeamSsml(ref),
+      );
       final awayTeamTTS = await textToSpeechService.getTtsNoFile(
-          text: selectedMatch.awayTeamSsml);
-
-      stopMessageRotation();
-      ref.read(isLoadingProvider.notifier).state = false;
+        text: selectedMatch.awayTeamSsml(ref),
+      );
 
       ref.read(azCharCountProvider.notifier).state +=
-          selectedMatch.ssml.length as int;
-      SettingsBox().azCharCount += selectedMatch.ssml.length as int;
+          selectedMatch.generateSsml(ref).length as int;
+      SettingsBox().azCharCount +=
+          selectedMatch.generateSsml(ref).length as int;
 
       // Play background music
 
       logger.d("[_handlePlayLineup] Starting background music");
 
-      await jingleManager.audioManager.playAudio(
-          AudioCategory.awayTeamJingle, ref,
-          shortFade: true, isBackgroundMusic: true);
+      // Find specific AwayJingle by configured file path or fallback
+      final awayJinglePath = SettingsBox().awayJingleFilePath;
+      AudioFile? awayJingle;
+
+      if (awayJinglePath.isNotEmpty) {
+        // Try to find the configured jingle by file path
+        awayJingle = jingleManager.audioManager.audioInstances
+            .where(
+              (instance) =>
+                  instance.audioCategory == AudioCategory.specialJingle &&
+                  instance.filePath == awayJinglePath,
+            )
+            .firstOrNull;
+      }
+
+      // Fallback to displayName search if configured jingle not found
+      if (awayJingle == null) {
+        logger.d(
+          "[_handlePlayLineup] Configured away jingle not found, falling back to displayName search",
+        );
+        awayJingle = jingleManager.audioManager.audioInstances
+            .where(
+              (instance) =>
+                  instance.audioCategory == AudioCategory.specialJingle &&
+                  instance.displayName == 'AwayJingle',
+            )
+            .firstOrNull;
+      }
+
+      if (awayJingle != null) {
+        await jingleManager.audioManager.playAudioFile(
+          awayJingle, // Play the specific AudioFile we found
+          ref,
+          shortFade: true,
+          isBackgroundMusic: true, // This will make it play as background music
+        );
+      }
 
       // wait for 10 seconds
 
@@ -223,14 +285,18 @@ class _LineupState extends ConsumerState<Lineup> {
       logger.d("[_handlePlayLineup] Playing welcome message");
 
       await jingleManager.audioManager.playBytesAndWait(
-          audio: welcomeTTS.audio.buffer.asUint8List(), ref: ref);
+        audio: welcomeTTS.audio.buffer.asUint8List(),
+        ref: ref,
+      );
 
       // Play away team lineup with background music
 
       logger.d("[_handlePlayLineup] Playing Away team background music");
 
       await jingleManager.audioManager.playBytesAndWait(
-          audio: awayTeamTTS.audio.buffer.asUint8List(), ref: ref);
+        audio: awayTeamTTS.audio.buffer.asUint8List(),
+        ref: ref,
+      );
 
       // wait for 10 seconds
 
@@ -243,8 +309,10 @@ class _LineupState extends ConsumerState<Lineup> {
 
       logger.d("[_handlePlayLineup] Fading out background music");
 
-      await jingleManager.audioManager
-          .fadeOutNoStop(ref, AudioChannel.channel1);
+      await jingleManager.audioManager.fadeOutNoStop(
+        ref,
+        AudioChannel.channel1,
+      );
 
       logger.d("[_handlePlayLineup] Stopping channel2");
 
@@ -254,9 +322,43 @@ class _LineupState extends ConsumerState<Lineup> {
 
       logger.d("[_handlePlayLineup] Playing Home team background music");
 
-      await jingleManager.audioManager.playAudio(
-          AudioCategory.homeTeamJingle, ref,
-          shortFade: true, isBackgroundMusic: true);
+      // Find specific HomeJingle by configured file path or fallback
+      final homeJinglePath = SettingsBox().homeJingleFilePath;
+      AudioFile? homeJingle;
+
+      if (homeJinglePath.isNotEmpty) {
+        // Try to find the configured jingle by file path
+        homeJingle = jingleManager.audioManager.audioInstances
+            .where(
+              (instance) =>
+                  instance.audioCategory == AudioCategory.specialJingle &&
+                  instance.filePath == homeJinglePath,
+            )
+            .firstOrNull;
+      }
+
+      // Fallback to displayName search if configured jingle not found
+      if (homeJingle == null) {
+        logger.d(
+          "[_handlePlayLineup] Configured home jingle not found, falling back to displayName search",
+        );
+        homeJingle = jingleManager.audioManager.audioInstances
+            .where(
+              (instance) =>
+                  instance.audioCategory == AudioCategory.specialJingle &&
+                  instance.displayName == 'HomeJingle',
+            )
+            .firstOrNull;
+      }
+
+      if (homeJingle != null) {
+        await jingleManager.audioManager.playAudioFile(
+          homeJingle, // Play the specific AudioFile we found
+          ref,
+          shortFade: true,
+          isBackgroundMusic: true, // This will make it play as background music
+        );
+      }
 
       // wait for 10 seconds
       logger.d("[_handlePlayLineup] Waiting 10 seconds");
@@ -268,7 +370,9 @@ class _LineupState extends ConsumerState<Lineup> {
       logger.d("[_handlePlayLineup] Playing Home team lineup");
 
       await jingleManager.audioManager.playBytesAndWait(
-          audio: homeTeamTTS.audio.buffer.asUint8List(), ref: ref);
+        audio: homeTeamTTS.audio.buffer.asUint8List(),
+        ref: ref,
+      );
 
       // wait for 10 seconds
 
@@ -286,12 +390,7 @@ class _LineupState extends ConsumerState<Lineup> {
       //     .playBytes(audio: speech.audio.buffer.asUint8List(), ref: ref);
     } catch (e) {
       logger.d("Error generating audio: $e");
-
       // You might want to show an error message to the user here
-    } finally {
-      // Stop message rotation and hide loading indicator
-      stopMessageRotation();
-      ref.read(isLoadingProvider.notifier).state = false;
     }
   }
 
