@@ -1,23 +1,28 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:soundboard/features/innebandy_api/data/class_lineup.dart';
-import 'package:soundboard/features/innebandy_api/data/class_match.dart';
+import 'package:soundboard/common/models/enum_goaltypes.dart';
+
+import 'package:soundboard/core/services/innebandy_api/domain/entities/lineup.dart';
+import 'package:soundboard/core/services/innebandy_api/domain/entities/match.dart';
 import 'package:soundboard/features/screen_home/presentation/lineup/classes/class_color_state_notifier.dart';
+import 'package:soundboard/features/screen_home/presentation/lineup/providers/manual_lineup_providers.dart';
 
 class LineupData extends ConsumerWidget {
   final double availableWidth;
   final double availableHeight;
 
-  const LineupData(
-      {required this.availableWidth, required this.availableHeight});
+  const LineupData({
+    required this.availableWidth,
+    required this.availableHeight,
+  });
 
   static const double _smallFontSize = 15.0;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final selectedMatch = ref.watch(selectedMatchProvider);
-    final selectedMatchLineup = ref.watch(lineupProvider);
+    final selectedMatchLineup = ref.watch(effectiveLineupProvider);
 
     if (selectedMatch.matchId == 0) {
       return const Center(child: Text("No Data"));
@@ -33,32 +38,44 @@ class LineupData extends ConsumerWidget {
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          _buildTeamColumn(context, ref, selectedMatch.homeTeam,
-              selectedMatchLineup.homeTeamPlayers, teamWidth),
+          _buildTeamColumn(
+            context,
+            ref,
+            selectedMatch.homeTeam,
+            selectedMatchLineup.homeTeamPlayers,
+            teamWidth,
+          ),
           Container(
             height: double.infinity,
             margin: const EdgeInsets.symmetric(horizontal: 4, vertical: 8),
             width: 1,
             color: Theme.of(context).colorScheme.onSurface,
           ),
-          _buildTeamColumn(context, ref, selectedMatch.awayTeam,
-              selectedMatchLineup.awayTeamPlayers, teamWidth),
+          _buildTeamColumn(
+            context,
+            ref,
+            selectedMatch.awayTeam,
+            selectedMatchLineup.awayTeamPlayers,
+            teamWidth,
+          ),
         ],
       ),
     );
   }
 
-  Widget _buildTeamColumn(BuildContext context, WidgetRef ref, String teamName,
-      List<TeamPlayer>? players, double width) {
+  Widget _buildTeamColumn(
+    BuildContext context,
+    WidgetRef ref,
+    String teamName,
+    List<TeamPlayer>? players,
+    double width,
+  ) {
     return SizedBox(
       width: width,
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          Text(
-            teamName,
-            style: Theme.of(context).textTheme.titleMedium,
-          ),
+          Text(teamName, style: Theme.of(context).textTheme.titleMedium),
           const SizedBox(height: 8),
           _buildPlayerList(context, ref, players),
         ],
@@ -67,11 +84,14 @@ class LineupData extends ConsumerWidget {
   }
 
   Widget _buildPlayerList(
-      BuildContext context, WidgetRef ref, List<TeamPlayer>? players) {
+    BuildContext context,
+    WidgetRef ref,
+    List<TeamPlayer>? players,
+  ) {
     if (players == null || players.isEmpty) {
       return const Center(child: Text("No players"));
     }
-    final buttonStates = ref.watch(playerStatesProvider);
+    final buttonStates = ref.watch(GoalTypeStatesProvider);
     final Map<String, String> positionMapping = {
       'Målvakt': 'MV',
       'Forward': 'F',
@@ -90,48 +110,48 @@ class LineupData extends ConsumerWidget {
       itemBuilder: (context, index) {
         final player = players[index];
         final playerId = '${player.shirtNo}-${player.name}';
-        var buttonState = buttonStates[playerId] ?? PlayerState.normal;
+        var buttonState = buttonStates[playerId] ?? GoalTypeState.normal;
         // Auto-highlight based on entered numbers
 
         Color getButtonColor() {
           switch (buttonState) {
-            case PlayerState.normal:
+            case GoalTypeState.normal:
               return Theme.of(context).colorScheme.surface;
-            case PlayerState.goal:
+            case GoalTypeState.goal:
               return Theme.of(context).colorScheme.primaryContainer;
-            case PlayerState.assist:
+            case GoalTypeState.assist:
               return Theme.of(context)
                   .colorScheme
                   .secondaryContainer; // You can change this to any color you prefer for long press
-            case PlayerState.penalty:
+            case GoalTypeState.penalty:
               return Theme.of(context).colorScheme.errorContainer;
           }
         }
 
         Color getTextColor() {
           switch (buttonState) {
-            case PlayerState.normal:
+            case GoalTypeState.normal:
               return Theme.of(context).colorScheme.onSurface;
-            case PlayerState.goal:
+            case GoalTypeState.goal:
               return Theme.of(context).colorScheme.onPrimaryContainer;
-            case PlayerState.assist:
+            case GoalTypeState.assist:
               return Theme.of(context)
                   .colorScheme
                   .onSecondaryContainer; // Adjust this based on your long press color
-            case PlayerState.penalty:
+            case GoalTypeState.penalty:
               return Theme.of(context).colorScheme.onErrorContainer;
           }
         }
 
         String getButtonStateText() {
           switch (buttonState) {
-            case PlayerState.normal:
+            case GoalTypeState.normal:
               return '';
-            case PlayerState.goal:
+            case GoalTypeState.goal:
               return 'MÅL';
-            case PlayerState.assist:
+            case GoalTypeState.assist:
               return 'ASSIST';
-            case PlayerState.penalty:
+            case GoalTypeState.penalty:
               return 'UTVISNING';
           }
         }
@@ -166,18 +186,20 @@ class LineupData extends ConsumerWidget {
                     backgroundColor: getButtonColor(),
                     foregroundColor: getTextColor(),
                     // padding: EdgeInsets.zero,
-                    padding:
-                        const EdgeInsets.symmetric(vertical: 0, horizontal: 16),
+                    padding: const EdgeInsets.symmetric(
+                      vertical: 0,
+                      horizontal: 16,
+                    ),
                     minimumSize: const Size(0, 30),
                   ),
                   onPressed: () {
                     ref
-                        .read(playerStatesProvider.notifier)
+                        .read(GoalTypeStatesProvider.notifier)
                         .setGoalState(playerId);
                   },
                   onLongPress: () {
                     ref
-                        .read(playerStatesProvider.notifier)
+                        .read(GoalTypeStatesProvider.notifier)
                         .setAssistState(playerId);
                   },
                   child: Row(
