@@ -7,11 +7,13 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:intl/intl.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:soundboard/features/cloud_text_to_speech/class_azure_region.dart';
-import 'package:soundboard/features/cloud_text_to_speech/providers.dart';
-import 'package:soundboard/features/scrolling/presentation/scroll_config.dart';
-import 'package:soundboard/properties.dart'; // Local file for handling soundboard properties.
-import 'package:soundboard/player_app.dart'; // Local main app file.
+import 'package:soundboard/core/services/cloud_text_to_speech/class_azure_region.dart';
+import 'package:soundboard/core/services/cloud_text_to_speech/providers.dart';
+import 'package:soundboard/core/utils/scroll_config.dart';
+import 'package:soundboard/core/properties.dart'; // Local file for handling soundboard properties.
+import 'package:soundboard/app.dart'; // Local main app file.
+import 'package:soundboard/features/screen_settings/data/class_slider_mappings_adapter.dart';
+import 'package:hive/hive.dart';
 
 // External Imports (sorted by package name length, shortest first)
 import 'package:easy_hive/easy_hive.dart'; // Package for easy integration of Hive, a lightweight and fast NoSQL database.
@@ -19,8 +21,9 @@ import 'package:easy_hive/easy_hive.dart'; // Package for easy integration of Hi
 
 // Flutter Imports
 import 'package:flutter/material.dart';
-import 'package:soundboard/theme_config.dart';
-import 'package:soundboard/utils/logger.dart';
+import 'package:soundboard/core/theme/theme_config.dart';
+import 'package:soundboard/core/utils/logger.dart';
+import 'package:soundboard/core/constants/globals.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -32,6 +35,9 @@ void main() async {
 
   Directory settingsDir = await getApplicationSupportDirectory();
   logger.d("AppSupportDir: ${settingsDir.path} ");
+
+  // Register the SliderMappingAdapter
+  Hive.registerAdapter(SliderMappingAdapter());
 
   await EasyBox.initialize(subDir: settingsDir.path);
   await SettingsBox().init();
@@ -86,7 +92,7 @@ class _SoundBoardState extends ConsumerState<SoundBoard> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      // navigatorKey: navigatorKey,
+      navigatorKey: navigatorKey,
       builder: (contex, child) {
         return ScrollConfiguration(
           behavior: MyCustomScrollBehavior(),
